+++ api/uisvc/log.go
@@ -1,4 +1,4 @@
-package processors
+package uisvc
 
 import (
 	"strconv"
