+++ api/uisvc/setup_test.go
@@ -1,4 +1,4 @@
-package test
+package uisvc
 
 import (
 	"testing"
